LayoutTests/imported/w3c/web-platform-tests/service-workers/cache-storage/serviceworker/cache-keys-attributes-for-service-worker.https-expected.txt


FAIL Request.IsReloadNavigation should persist. assert_equals: expected "original: false, stored: false" but got "original: undefined, stored: undefined"
FAIL Request.IsHistoryNavigation should persist. assert_equals: expected "original: false, stored: false" but got "original: undefined, stored: undefined"
